+++ warphog/cores.py
@@ -44,14 +44,9 @@ class CPUPreWarpCore(WarpCore):
 
     def __init__(self, seq_block, alphabet):
         super().__init__(seq_block, alphabet)
-        #self.data_block = self._make_data_block(seq_block)
 
         self.kernel = KERNELS["python"]()
         self.kernel.prepare_kernel(alphabet=alphabet)
-
-    def _make_data_block(self, seq_block):
-        msa_char_block = np.fromstring("".join(seq_block).encode(), dtype=np.byte)
-        return msa_char_block
 
     def put_d(self, d):
         self.d = d
@@ -81,7 +76,7 @@ class GPUWarpCore(WarpCore):
         self.d.get(d)
 
     def _make_data_block(self, seq_block):
-        msa_char_block = np.frombuffer("".join(seq_block).encode(), dtype=np.byte)
+        msa_char_block = np.frombuffer(b''.join(seq_block), dtype=np.byte)
         msa_gpu = gpuarray.to_gpu(msa_char_block)
         return msa_gpu